% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotation_compass}
\alias{annotation_compass}
\title{annotation_compass: taken from 
https://stackoverflow.com/questions/47916307/ggplot2-specify-geom-text-position-by-conventional-top-bottom-left-rig}
\usage{
annotation_compass(
  label,
  position = c("N", "NE", "E", "SE", "S", "SW", "W", "NW"),
  padding = grid::unit(c(0.5, 0.5), "line"),
  ...
)
}
\arguments{
\item{label}{label}

\item{position}{position}
}
\description{
annotation_compass: taken from 
https://stackoverflow.com/questions/47916307/ggplot2-specify-geom-text-position-by-conventional-top-bottom-left-rig
}
