% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{ceiling_dec: taken from 
https://stackoverflow.com/questions/35807523/r-decimal-ceiling}
\usage{
ceiling_dec(x, level = 1)
}
\arguments{
\item{label}{label}

\item{position}{position}
}
\description{
ceiling_dec: taken from 
https://stackoverflow.com/questions/35807523/r-decimal-ceiling
}
