% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-general-functions.R
\name{closest_date}
\alias{closest_date}
\title{This finds the closest date to a given date that exists in a dataset.}
\usage{
closest_date(data, date, type = "all", force_val = T)
}
\arguments{
\item{data}{A data frame that includes a column `timestamp`}

\item{date}{The date you want to match`}

\item{type}{Default "all", which returns closest date overall. Other options are "before", "after", "beforeequal", and "afterequal"`}

\item{force_val}{Default True. If True, then the function will return the "all" type if there is no date given for the type}
}
\description{
This finds the closest date to a given date that exists in a dataset.
}
\examples{
closest_date(data = df, date = "2020-01-01", type = "all", force_val = T)
}
