% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorschemer}
\alias{colorschemer}
\title{Colorscheme}
\usage{
colorschemer(col)
}
\arguments{
\item{col}{A data frame that includes a column `timestamp`}
}
\description{
Colorscheme
}
\examples{
colorschemer("red")
}
