% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeline.R
\name{create_time_batches}
\alias{create_time_batches}
\title{create_time_batches}
\usage{
create_time_batches(start, end, year_batch)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}
}
\description{
create_time_batches
}
\examples{
getTimelinesForHealth()
}
