% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spikes-onegeo.R
\name{get_rawcounts}
\alias{get_rawcounts}
\title{get_rawcounts using estimates from Comscore}
\usage{
get_rawcounts(
  df,
  month = 2,
  pct_desktop = 0.35,
  geo = "US",
  interrupt = "2020-03-01",
  qf_denominator = 1e+07,
  endperiod = T
)
}
\arguments{
\item{df}{The dataframe as outputted by \code{run_arima}.}

\item{month}{1 or 2, 1 is the earlier month and 2 is the later. Default 2}

\item{pct_desktop}{Assumption of the percentage of searches that are executed on Desktops. Default 0.35}

\item{geo}{Location}

\item{interrupt}{The date of the interruption (should be the same as \code{run_arima})}

\item{qf_denominator}{Denominator for fractions given in gtrendspy. Default 10M.}

\item{endperiod}{Compute raw counts from the interrupt to an enddate}
}
\description{
get_rawcounts using estimates from Comscore
}
\examples{
get_rawcounts <- function(
  df,
  month = 2,
  pct_desktop = 0.35,
  geo = "US",
  interrupt = "2020-03-01",
  qf_denominator = 10000000,
  endperiod = T
  )
}
