% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeline.R
\name{getTimelinesForHealth}
\alias{getTimelinesForHealth}
\title{getTimelineForHealth}
\usage{
getTimelinesForHealth(
  batch_size = 1,
  year_batch = "1 year",
  time.startDate = "2018-06-15",
  time.endDate = "2020-01-01",
  timelineResolutions = c("day"),
  terms = c("summer + winter + fall + spring", "cat + cat food + dog + dog food"),
  names = c("seasons", "pets"),
  geoRestriction.regions = c("US-NY", "US-CA"),
  geoRestriction.countries = c("GB", "US"),
  geoRestriction.dmas = c(),
  output_directory = "../output"
)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}
}
\description{
getTimelineForHealth
}
\examples{
getTimelinesForHealth()
}
