% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spikes-onegeo.R
\name{line_plot}
\alias{line_plot}
\title{line_plot: Creates a simple line plot of searches over time}
\usage{
line_plot(
  df,
  geo = "US",
  beginplot = T,
  endplot = T,
  interrupt,
  linelabel = "Interruption",
  linelabelpos = 0.02,
  title = NULL,
  xlab = "Date",
  ylab = "Query Fraction\\n(Per 10 Million Searches)",
  lbreak = "1 year",
  ylim = NULL,
  hicol = NA,
  locol = NA,
  nucol = NA,
  opcol = NA,
  colorscheme = "red",
  xfmt = date_format("\%b \%Y"),
  lwd = 0.3,
  extend = F,
  width = 6,
  height = 3,
  save = T,
  outfn
)
}
\arguments{
\item{df}{The dataframe as outputted by \code{run_arima}.}

\item{geo}{The column name of the geography of searches that you want.}

\item{beginplot}{The date you want the plot to start. Default is beginning of the time series.}

\item{endplot}{The date you want the plot to end. Default is end of the time series.}

\item{interrupt}{The date of the interruption (should be the same as \code{run_arima})}

\item{linelabel}{Label next to the vertical interruption line}

\item{title}{The title of the figure. The default is no title.}

\item{xlab}{The label for the x axis}

\item{ylab}{The label for the y axis}

\item{lbreak}{The distance between tick marks in the x axis, i.e., "one year" or "3 month"}

\item{lwd}{Line width}

\item{width}{Width of the plot in inches}

\item{height}{Height of the plot in inches}

\item{save}{Default is True. If False, the plot is not saved.}

\item{outfn}{Where to save the plot.}
}
\description{
line_plot: Creates a simple line plot of searches over time
}
\examples{
line_plot(
  df,
  beginplot = T,
  endplot = T,
  interrupt = ymd("2019-12-19")
  linelabel = "Interruption",
  title = NULL,
  xlab = "Date",
  ylab = "Query Fraction (Per 10 Million Searches)",
  lbreak = "1 year",
  lwd = 0.3,
  width = 6,
  height = 3,
  save = T,
  outfn = './output/panA.png'
)
}
