% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multiterms.R
\name{multi_term_arima}
\alias{multi_term_arima}
\title{multi_term_arima}
\usage{
multi_term_arima(
  input_dir = "C:/Users/tcapu/Google Drive/modules/gtrendR/READMEcode/input",
  geo = "US",
  terms_to_use = NA,
  timeframe_to_use = NA,
  beginperiod = T,
  preperiod = 90,
  endperiod = T,
  interrupt = "2020-03-01",
  bootstrap = T,
  bootnum = 1000,
  na = "kalman",
  kalman = F,
  include_data = T,
  linear = F,
  min0 = F,
  logit = T,
  logit_a = 0,
  logit_b = 100
)
}
\arguments{
\item{beginperiod}{How far back you want the "pre" period to go}

\item{preperiod}{This creates a beginperiod but with a number of days instead of a date}

\item{endperiod}{How far after the interruption you want to go}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}

\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{scaletitle}{Title of the scale}

\item{scalelimits}{vector of two values for min and max for the scale}

\item{linecol}{Line color}

\item{lowcol}{Color for low values}

\item{midcol}{Color for mid values}

\item{highcol}{Color for high values}

\item{save}{Default is True, If False, don't save}

\item{width}{Width of file in inches}

\item{height}{Height of file in inches}

\item{outfn}{Output filename}
}
\description{
multi_term_arima
}
\examples{
multiterms <- multi_term_arima(

  ## A folder containing all of your gtrends data
  input_dir = "./input",

  ## Which data to use
  geo = "US", # Geography you want to use
  terms_to_use = NA, # Terms you'd like to analyze. If NA then all terms
  timeframe_to_use = NA, # Only analyze data with filenames that contain a certain timeframe. If NA then all timeframes


  ## Parameters of time periods
  beginperiod = T, # Beginning of the before period, if T then beginning of data
  preperiod = 90, # If beginperiod is logical, preperiod is the number of days before interrupt to include in before period
  endperiod = T, # End of the end period, if T then end of data
  interrupt = "2020-03-01", # Date for interruption, splitting before and after periods


  ## Analytical arguments
  bootstrap = T, # Bootstrap CIs
  bootnum = 1000, # Number of bootstraps
  kalman = T # If T, impute with Kalman
)
}
