% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multiterms.R
\name{multiterm_barplot}
\alias{multiterm_barplot}
\title{multiterm_barplot: Use the data from multi_term_arima to create a barplot}
\usage{
multiterm_barplot(
  df,
  label_df = NA,
  title = NULL,
  xlab = "Terms",
  ylab = "Greater than Expected (\%)",
  ylim = NULL,
  space = 0.8,
  colorscheme = "blue",
  hicol = NA,
  save = T,
  outfn = "./output/panG.png",
  width = 6,
  height = 3
)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{save}{Default is True, If False, don't save}

\item{outfn}{Output filename}

\item{width}{Width of file in inches}

\item{height}{Height of file in inches}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}

\item{beginperiod}{How far back you want the "pre" period to go}

\item{preperiod}{This creates a beginperiod but with a number of days instead of a date}

\item{endperiod}{How far after the interruption you want to go}

\item{scaletitle}{Title of the scale}

\item{scalelimits}{vector of two values for min and max for the scale}

\item{linecol}{Line color}

\item{lowcol}{Color for low values}

\item{midcol}{Color for mid values}

\item{highcol}{Color for high values}
}
\description{
multiterm_barplot: Use the data from multi_term_arima to create a barplot
}
\examples{
panG <- multiterm_barplot(
  df = multiterms,

  ## Graphing Parameters
  title = NULL, # If NULL, no Title
  xlab = "Terms", # x axis label
  label_df = NA, # Use a two-column dataframe to label the barplot x axis
  ylab = "Greater than Expected (\%)", # y axis label
  space = 0.8, # space between bars

  ## Set a colorscheme
  colorscheme = "blue",  # Color schemes set in this package "red", 'blue" or "jamaim"

  # ... customize any color using these
  hicol = NA, # Color of bars

  ## Saving arguments
  save = T, # If T, save plot
  outfn = './output/panG.png', # Location to save plot
  width = 6, # Width in inches
  height = 3 # Height in inches
)
}
