% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multiterms.R
\name{multiterm_spaghetti}
\alias{multiterm_spaghetti}
\title{multiterm_spaghetti: Use the data from multi_term_arima to create a spaghetti plot}
\usage{
multiterm_spaghetti(
  multiterm_list,
  interrupt = "2020-03-01",
  terms_to_use = NA,
  beginplot = "2020-03-01",
  endplot = "2020-04-03",
  title = NULL,
  xlab = "Date",
  lbreak = "1 week",
  xfmt = date_format("\%b-\%d"),
  ylab = "Query Fraction\\n(Per 10 Million Searches)",
  lwd = 1,
  spaghettilwd = 0.2,
  ylim = c(NA, NA),
  spaghettialpha = 0.25,
  colorscheme = "blue",
  hicol = NA,
  locol = NA,
  save = T,
  outfn = "./output/panE.png",
  width = 6,
  height = 3
)
}
\arguments{
\item{multiterm_list}{# from multi_term_arima}
}
\description{
multiterm_spaghetti: Use the data from multi_term_arima to create a spaghetti plot
}
