% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spikes-onegeo.R
\name{run_arima}
\alias{run_arima}
\title{Run this first. It applies the ARIMA to predict future values and then produces a dataset with the actual and fitted values.}
\usage{
run_arima(
  df,
  interrupt,
  begin = T,
  end = T,
  geo = "US",
  kalman = F,
  bootstrap = F,
  bootnum = 1000
)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}

\item{geo}{The column name of the geography of searches that you want.}

\item{polycolor}{What color do you want the polygon to be?.}
}
\description{
Run this first. It applies the ARIMA to predict future values and then produces a dataset with the actual and fitted values.
}
\examples{
run_arima(df = data, interrupt = ymd("2019-12-19"), geo = "US")
}
