% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimelinesForHealth.R
\name{seqlast}
\alias{seqlast}
\title{seqlast}
\usage{
seqlast(from, to, by)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}
}
\description{
seqlast
}
\examples{
getTimelinesForHealth()
}
