% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multigeo.R
\name{state_arima}
\alias{state_arima}
\title{state_arima: Run this first to do ARIMA by state functions}
\usage{
state_arima(data, begin = T, end = T, interrupt = "2020-03-01")
}
\arguments{
\item{data}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{begin}{Use data after this date to make a prediction}

\item{end}{Predict up to this date}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}
}
\description{
state_arima: Run this first to do ARIMA by state functions
}
\examples{
state_list <- state_arima(
  data = read.csv("./temp/data.csv", header = T, stringsAsFactor = F),
  interrupt = "2019-03-01"
)
}
