% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multigeo.R
\name{state_arima_pctdiff}
\alias{state_arima_pctdiff}
\title{state_arima_pctdiff: Run this first to do ARIMA by state functions}
\usage{
state_arima_pctdiff(
  state_arima_list,
  linecol = "gray",
  scaletitle = "\% Diff.\\nin Searches",
  lowcol = "white",
  midcol = NULL,
  highcol = "dodgerblue4",
  save = T,
  width = 6,
  height = 4,
  outfn = "./output/fig.png"
)
}
\arguments{
\item{state_arima_list}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{linecol}{Color of the state boundaries}

\item{scaletitle}{Title on the scale}

\item{lowcol}{Color at lower end of scale}

\item{midcol}{Color at mid of scale}

\item{highcol}{Color of high end of scale}

\item{save}{default T, if false does not save}

\item{width}{width in inches}

\item{height}{height in inches}

\item{outfn}{output filename}
}
\description{
state_arima_pctdiff: Run this first to do ARIMA by state functions
}
\examples{
state_arima_pctdiff(
  state_list,
  linecol = "gray",
  scaletitle = "\% Diff.\nin Searches",
  lowcol = "white",
  midcol = NULL,
  highcol = "dodgerblue4",
  save = T,
  width = 6,
  height = 4,
  outfn = "./output/fig.png"
)
}
