% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multigeo.R
\name{state_arima_spaghetti}
\alias{state_arima_spaghetti}
\title{state_arima_spaghetti: Use the output of state_arima to create a state-level spaghetti plot of differences between the ARIMA model and actual searches}
\usage{
state_arima_spaghetti(
  state_arima_list,
  beginplot,
  endplot,
  interrupt = NA,
  ylim = NULL,
  title = NULL,
  xlab = "Date",
  ylab = "Actual Versus Model-Fitted\\nSearch Queries (\% Diff.)",
  linelabel = "Interruption",
  lbreak = "1 week",
  lwd = 0.4,
  xfmt = date_format("\%d \%b"),
  states_with_labels = c("CA", "NY", "US", "IL", "TX"),
  states_to_exclude = c(),
  hicol = NA,
  locol = NA,
  nucol = NA,
  opcol = NA,
  colorscheme = "red",
  spaghettialpha = 0.25,
  extend = F,
  save = T,
  width = 6,
  height = 4,
  outfn = "./output/fig.png"
)
}
\arguments{
\item{state_arima_list}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{beginplot}{When the plot begins}

\item{endplot}{When the plot ends}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}

\item{ylim}{A 2-item vector for the limits of the y axis}

\item{title}{Title of the figure}

\item{xlab}{x-axis label, default is "Date"}

\item{ylab}{y-axis label, default is "Actual Versus Model-Fitted Search Queries (% Diff.)"}

\item{linelabel}{label next to interruption line}

\item{lbreak}{space between tick marks on x axis}

\item{lwd}{line width}

\item{xfmt}{date format for x axis}

\item{states_with_labels}{vector of states that you want to have labels}

\item{states_to_exclude}{vector of states not to include in the figure}

\item{save}{default T, if false does not save}

\item{width}{width in inches}

\item{height}{height in inches}

\item{outfn}{output filename}
}
\description{
state_arima_spaghetti: Use the output of state_arima to create a state-level spaghetti plot of differences between the ARIMA model and actual searches
}
\examples{
state_arima_spaghetti(
  state_list,
  interrupt = "2019-03-01",
  title = NULL,
  xlab = "Date",
  ylab = "Actual Versus Model-Fitted\nSearch Queries (\% Diff.)",
  data = read.csv("./temp/data.csv", header = T, stringsAsFactor = F),
  linelabel = "COVID-19\nOutbreak",
  lbreak = "1 week",
  lwd = 0.4,
  beginplot = ymd("2019-03-01")-(7*1),
  endplot = ymd("2019-03-18"),
  xfmt = date_format("\%d \%b"),
  states_with_labels = c("CA", "NY", "US", "IA"),
  states_to_exclude = c("IA"),
  save = T,
  width = 6,
  height = 4,
  outfn = "./output/fig.png"
)
}
