% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-spike-multigeo.R
\name{state_pct_change}
\alias{state_pct_change}
\title{pct_change_state: Create a map of states based upon the percentage increase in searches before or after the interruption}
\usage{
state_pct_change(
  df,
  interrupt,
  beginperiod = NA,
  preperiod,
  endperiod,
  scaletitle = "\% Increase\\nin Searches",
  scalelimits = NULL,
  linecol = "gray",
  lowcol = "#8BADE1",
  midcol = NULL,
  highcol = "#EB2836",
  save = T,
  width = 6,
  height = 4,
  outfn = "./output/fig.png",
  bootnum = 1000,
  alpha = 0.05,
  return_df = T
)
}
\arguments{
\item{df}{A dataframe including time as \code{timestamp} and searches for your given geography in one column.}

\item{interrupt}{The date where things change. ARIMA will be predicted on all days before the interrupt.}

\item{beginperiod}{How far back you want the "pre" period to go}

\item{preperiod}{This creates a beginperiod but with a number of days instead of a date}

\item{endperiod}{How far after the interruption you want to go}

\item{scaletitle}{Title of the scale}

\item{scalelimits}{vector of two values for min and max for the scale}

\item{linecol}{Line color}

\item{lowcol}{Color for low values}

\item{midcol}{Color for mid values}

\item{highcol}{Color for high values}

\item{save}{Default is True, If False, don't save}

\item{width}{Width of file in inches}

\item{height}{Height of file in inches}

\item{outfn}{Output filename}
}
\description{
pct_change_state: Create a map of states based upon the percentage increase in searches before or after the interruption
}
\examples{
pct_change_state(
  df = read.csv("./temp/data.csv", header = T, stringsAsFactor = F),
  interrupt = "2020-03-01",
  beginperiod = NA,
  preperiod = 90,
  endperiod = "2020-03-23",
  scaletitle = "Pct. Increase\nin Searches",
  linecol = "gray",
  lowcol = "red",
  midcol = "white",
  highcol = "dodgerblue4"
)
}
